# Configure Style/NumericPredicate for performance