--- a/.rubocop.yml
+++ b/.rubocop.yml
@@ -59,3 +59,6 @@ Style:
 Style/NumericLiterals:
   Exclude:
     - 'lib/**/*'
+
+Style/NumericPredicate:
+  EnforcedStyle: comparison