--- a/lib/axlsx/content_type/abstract_content_type.rb
+++ b/lib/axlsx/content_type/abstract_content_type.rb
@@ -28,7 +28,7 @@ module Axlsx
     def to_xml_string(node_name = '', str = +'')
       str << '<' << node_name << ' '
       Axlsx.instance_values_for(self).each_with_index do |key_value, index|
-        str << ' ' unless index.zero?
+        str << ' ' unless index == 0
         str << Axlsx.camel(key_value.first) << '="' << key_value.last.to_s << '"'
       end
       str << '/>'