--- a/lib/axlsx/rels/relationship.rb
+++ b/lib/axlsx/rels/relationship.rb
@@ -116,7 +116,7 @@ module Axlsx
       h = Axlsx.instance_values_for(self).reject { |k, _| k == "source_obj" }
       str << '<Relationship '
       h.each_with_index do |key_value, index|
-        str << ' ' unless index.zero?
+        str << ' ' unless index == 0
         str << key_value.first.to_s << '="' << Axlsx.coder.encode(key_value.last.to_s) << '"'
       end
       str << '/>'