Switch from using predicate methods like `v.zero?` to direct comparisons
like `v == 0` for performance optimization.

Although this may not yield a visible improvement with the current
implementation, using `== 0` instead of `.zero?` in a method for cell
rendering could lead to significant performance gains in the future.

Benchmarks show that comparisons are faster than predicates across
multiple Ruby versions:

```
Comparison (Ruby 2.6 to 3.2, v = 0):
              v == 0: 21479329.3 i/s
             v.zero?: 17979885.4 i/s - 1.19x  (± 0.00) slower

Comparison (Ruby 3.3, v = 0):
              v == 0: 23652215.7 i/s
             v.zero?: 21843174.0 i/s - 1.08x  (± 0.00) slower

Comparison (Ruby 2.6 to 3.3, v = 1):
              v == 0: 23227474.2 i/s
             v.zero?: 21675200.9 i/s - 1.07x  (± 0.00) slower
```

---

Hello, 

~I'm having slightly and consistently better results on this branch. It is more noticeable on x64 Ruby. It depends on Ruby version and it's about 6% on 2.6.10 and 2% on 3.3.5~

~Can you try `test/benchmark.rb` (several times) to see if you have the same results?~

I can't confirm overall performance improvements, because this method is only invoked 29 times in benchmarks. However, if `,zero?` will be added to a method used to render cells, it would be a problem