@tagliala 2024-09-08
Failure is not related. Ref: simplecov-ruby/simplecov-html#144